// Team-5 Frontend/src/app/Model/Books.ts
export class Books{
    bookId: string ="";
    sellerId:number=0;
    bookName: string="";
    price: number=0;
    authorName: string="";
    bookDetails: string="";
    createdAt: string="";
    updatedAt:string="";
    status: string="Selling";
    image: string="";
    bookLanguage:string="";
	bookCondition:string="";
    pickUpAddress: string="";
    category:string="";



}

// //	private LocalDateTime createdDateAndTime;
// //	private LocalDateTime updatedDateAndTime;
// 	private String status;
// 	private String image;
// 	private String pickUpAddress;